+++ xeofs/validation/__init__.py
@@ -1,1 +1,1 @@
-from ._base_bootstrapper import EOFBootstrapper
+from .bootstrapper import EOFBootstrapper
